+++ clients/derivatives-trading-portfolio-margin/tests/unit/rest-api/AccountApiTest.test.ts
@@ -2404,7 +2404,7 @@ describe('AccountApi', () => {
                         asset: 'USDT',
                         info: 'TRANSFER',
                         time: 1570608000000,
-                        tranId: '9689322392',
+                        tranId: 9689322392,
                         tradeId: '',
                     },
                     {
@@ -2414,7 +2414,7 @@ describe('AccountApi', () => {
                         asset: 'USDT',
                         info: 'COMMISSION',
                         time: 1570636800000,
-                        tranId: '9689322392',
+                        tranId: 9689322392,
                         tradeId: '2059192',
                     },
                 ])
@@ -2454,7 +2454,7 @@ describe('AccountApi', () => {
                         asset: 'USDT',
                         info: 'TRANSFER',
                         time: 1570608000000,
-                        tranId: '9689322392',
+                        tranId: 9689322392,
                         tradeId: '',
                     },
                     {
@@ -2464,7 +2464,7 @@ describe('AccountApi', () => {
                         asset: 'USDT',
                         info: 'COMMISSION',
                         time: 1570636800000,
-                        tranId: '9689322392',
+                        tranId: 9689322392,
                         tradeId: '2059192',
                     },
                 ])